projects/🌌 CosmicFinance_project/requirements.txt
setuptools>=68.0.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary>=2.9.9
python-decouple==3.8
django-filter==23.3
pytest==7.4.3
pytest-django==4.7.0

